Import create student and teacher to `app.rb`
