+++ app.rb
@@ -2,7 +2,6 @@ require_relative 'student'
 require_relative 'teacher'
 require_relative 'book'
 require_relative 'rental'
-require_relative 'add_person'
 
 class App
   def initialize
@@ -17,6 +16,40 @@ class App
 
   def list_people
     @people.each { |person| puts "[#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}" }
+  end
+
+  def create_student
+    print 'Age: '
+    age = gets.chomp.to_i
+    print 'Name: '
+    name = gets.chomp
+    print 'Has parent permission? [Y/N]: '
+    permission = gets.chomp.downcase
+
+    case permission
+    when 'y'
+      student = Student.new(age, nil, name, parent_permission: true)
+      @people << student
+    when 'n'
+      student = Student.new(age, nil, name, parent_permission: false)
+      @people << student
+    else
+      'You have entered an invalid option'
+    end
+
+    puts 'You have successfully registered a Student'
+  end
+
+  def create_teacher
+    print "Teacher's name: "
+    name = gets.chomp
+    print "Teacher's age: "
+    age = gets.chomp.to_i
+    print "Teacher's specialization: "
+    specialization = gets.chomp
+    teacher = Teacher.new(age, specialization, name, parent_permission: true)
+    @people << teacher
+    puts 'You have successfully registered a Teacher'
   end
 
   def create_person
